requirements.txt
streamlit==1.54.0
ultralytics==8.4.14
opencv-python-headless>=4.10,<5
pandas==2.3.3
plotly==6.5.2
Pillow==12.1.1
streamlit-folium==0.26.2
folium==0.20.0
reportlab==4.4.10
numpy>=2.0,<3
